+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 0.2.0 (????-??-??)
 
  * Update `memfd` to 0.3
+ * Switch from `failure` to `thiserror`